Refactored Bitcoin APIs
