+++ foundations/api/resources/transaction_input.py
@@ -3,33 +3,15 @@
 # Usecase: Bitcoin transaction input API
 # Functionality: GET
 
-from common.utils import serialize_transaction_input, serialize_address
 from flask_restful import Resource
 from models.models import Input as  TransactionInput
 from models.models import db_session, Output, OutputAddress, Address
 from models.response_codes import ResponseCodes
 from models.response_codes import ResponseDescriptions
+from utils.serialize import serialize_transaction_input, serialize_address
+from utils.validate import validate_transaction_ids
 from webargs import fields
 from webargs.flaskparser import use_kwargs
-
-
-def ValidateTransactionIds(transaction_ids):
-    """
-    Method to validate transaction ids
-    :param transaction_ids:
-    """
-    validationErrorList = []
-    if len(transaction_ids) == 0:
-        validationErrorList.append({"ErrorMessage": ResponseDescriptions.TransactionIdsInputMissing.value})
-    if len(transaction_ids) > 10:
-        validationErrorList.append({"ErrorMessage": ResponseDescriptions.NumberOfTransactionIdsLimitExceeded.value})
-    if len(transaction_ids) > 0:
-        for transaction_id in transaction_ids:
-            if transaction_id <= 0:
-                validationErrorList.append(
-                    {"ErrorMessage": ResponseDescriptions.InvalidTransactionIdsInputValues.value})
-                break
-    return validationErrorList
 
 
 class TransactionInputEndpoint(Resource):
@@ -49,7 +31,7 @@ class TransactionInputEndpoint(Resource):
             request = {"transaction_ids": transaction_ids}
             response = {}
             # Validate User Input
-            validations_result = ValidateTransactionIds(transaction_ids)
+            validations_result = validate_transaction_ids(transaction_ids)
             if validations_result is not None and len(validations_result) > 0:
                 response = {"ResponseCode": ResponseCodes.InvalidRequestParameter.value,
                             "ResponseDesc": ResponseCodes.InvalidRequestParameter.name,
